requirements.txt
Deprecated==1.2.14
googleapis-common-protos==1.65.0
grpcio==1.67.1
importlib_metadata==8.4.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
protobuf==4.25.5
typing_extensions==4.12.2
wrapt==1.16.0
zipp==3.20.2
